exporter: Check /usr/sbin for ser2net
On Debian, the ser2net binary is installed at /usr/sbin/ser2net and it
can be run by any user. Since this isn't on the users PATH, we should
attempt to use it when failing to find a ser2net binary, before falling
back to /usr/bin/ser2net.

Signed-off-by: Thomas Preston <thomas.preston@codethink.co.uk>